requirements.txt
mss==9.0.1
Pillow==10.1.0
sounddevice==0.4.6
numpy==1.26.2
moviepy==1.0.3
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
